feat: add dprint and lychee tools with configs
Dprint configurations only includes the plugin for markdown, more
plugins have to be added on any new languages which need formatting.

Lychee's default config file was added, the only non default
configuration is the list of accepted status codes for URLs,
Informational codes were added as to assert the existence of the
resource, and the 403 code was added as some pages such as leetcode,
return such code to bots.